requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=1.4.0=pypi_0
accelerate=0.18.0=pypi_0
aiofiles=22.1.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
aiosqlite=0.19.0=pypi_0
albumentations=1.3.0=pypi_0
anyio=3.6.2=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
babel=2.12.1=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.12.2=pypi_0
bleach=6.0.0=pypi_0
ca-certificates=2022.12.7=ha878542_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=3.1.0=pypi_0
clip=1.0=pypi_0
cmake=3.26.3=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.1.3=pyhd8ed1ab_0
contourpy=1.0.7=pypi_0
cudatoolkit=11.8.0=h37601d7_11
cycler=0.11.0=pypi_0
cython=0.29.34=pypi_0
datasets=2.11.0=pypi_0
debugpy=1.6.7=py38h8dc9893_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
diffusers=0.15.1=pypi_0
dill=0.3.6=pypi_0
dlib=19.24.1=pypi_0
easydict=1.10=pypi_0
executing=1.2.0=pyhd8ed1ab_0
facenet-pytorch=2.5.3=pypi_0
fastjsonschema=2.16.3=pypi_0
filelock=3.11.0=pypi_0
flatbuffers=23.3.3=pypi_0
fonttools=4.39.3=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.4.0=pypi_0
ftfy=6.1.1=pypi_0
gast=0.4.0=pypi_0
google-auth=2.17.3=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.54.0=pypi_0
h5py=3.8.0=pypi_0
huggingface-hub=0.13.4=pypi_0
humanfriendly=10.0=pypi_0
idna=3.4=pypi_0
imageio=2.27.0=pypi_0
importlib-metadata=6.5.0=pyha770c72_0
importlib-resources=5.12.0=pypi_0
importlib_metadata=6.5.0=hd8ed1ab_0
insightface=0.7.3=pypi_0
ipykernel=6.22.0=pyh210e3f2_0
ipython=8.12.0=pyh41d4057_0
ipython-genutils=0.2.0=pypi_0
isoduration=20.11.0=pypi_0
jax=0.4.8=pypi_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
json5=0.9.11=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=2.5.0=pypi_0
jupyter-server-fileid=0.9.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyter-server-ydoc=0.8.0=pypi_0
jupyter-ydoc=0.2.4=pypi_0
jupyter_client=8.2.0=pyhd8ed1ab_0
jupyter_core=5.3.0=py38h578d9bd_0
jupyterlab=3.6.3=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.22.1=pypi_0
keras=2.12.0=pypi_0
kiwisolver=1.4.4=pypi_0
lazy-loader=0.2=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.0=pypi_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=12.2.0=h65d4601_19
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=12.2.0=h46fd767_19
lit=16.0.1=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=2.0.5=pypi_0
ml-dtypes=0.1.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
nbclassic=0.5.5=pypi_0
nbclient=0.7.3=pypi_0
nbconvert=7.3.1=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pypi_0
notebook=6.5.4=pypi_0
notebook-shim=0.2.2=pypi_0
numpy=1.23.5=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.2=pypi_0
onnx=1.13.1=pypi_0
onnxruntime-gpu=1.14.1=pypi_0
opencv-python=4.7.0.72=pypi_0
opencv-python-headless=4.7.0.72=pypi_0
openssl=1.1.1t=h0b41bf4_0
opt-einsum=3.3.0=pypi_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.0=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.5.0=pypi_0
pip=23.0.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.2.0=pyhd8ed1ab_0
prettytable=3.7.0=pypi_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=3.20.3=pypi_0
psutil=5.9.5=py38h1de0b5d_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pygments=2.15.1=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.8.16=h7a1cb2a_3
python-dateutil=2.8.2=pyhd8ed1ab_0
python-json-logger=2.0.7=pypi_0
python_abi=3.8=2_cp38
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.2=py38he24dcef_0
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
regex=2023.3.23=pypi_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.9.1=pypi_0
seaborn=0.12.2=pypi_0
semdiffusers=1.0.0=pypi_0
send2trash=1.8.0=pypi_0
setuptools=66.0.0=py38h06a4308_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=pypi_0
tensorboard=2.12.2=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.12.0=pypi_0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.32.0=pypi_0
termcolor=2.2.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.4.12=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
tomli=2.0.1=pypi_0
torch=2.0.0=pypi_0
torchvision=0.15.1=pypi_0
tornado=6.3=py38h1de0b5d_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
transformers=4.28.1=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2023.3=pypi_0
uri-template=1.2.0=pypi_0
urllib3=1.26.15=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py38h06a4308_0
wrapt=1.14.1=pypi_0
xxhash=3.2.0=pypi_0
xz=5.2.10=h5eee18b_1
y-py=0.5.9=pypi_0
yarl=1.8.2=pypi_0
ypy-websocket=0.8.2=pypi_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
